Fix Fillable's __getattr__ method
We should retrieve an AttributeError whenever we have already retrieved
the data from the server and we don't have any attribute with the
required name.

Which is whenever __getattr__ is called and we have already retrieved
data from the server and have filled "self" with new attributes for that
data.