Pass generation info to delete object
When deleting an object from a bucket we may want to delete a specific
version of the object, and for that we need to pass generation argument.

With this patch we apss it to the delete method and use it; before we
were only passing it but not using it.