Make list a generic method
Before this patch Project and Bucket had its own list implementation
that was basically identical.

This patch introduces an ABC called Listable and abstract common code.

While this change may seem a little convoluted at the moment it will be
helpful in we decide to create most of the classes dynamically using
discovery.