Add support for delimiter in listings
GCS support setting a delimiter so we can browse objects in a bucket as
if we were browsing filesystem directories.

This patch adds support to retrieve prefixes field on a listing and
turns them into Prefix objects that can then be listed themselves.

This means that now when we list a bucket if we set delimiter optional
argument we can receive a mix of Object and Prefix instances.

For now Prefix instances are only good to list their contents.