Add metadata and timeDeleted values for Objects
This patch adds default values for attributes that only exists in Object
instances under specific circumstances.

metadata will not exist if there is no metadata in the object, so we
initialize it to an empty dictionary.

timeDeleted will only exist on object versions that have been deleted.
We initialize it to None.