Allow formatting url in _request method
The "_request" method accepts a "url" parameter, but it didn't allow
formatting it automatically, and it was used as is.

This patch adds the possibility, enabled by default, to format the url
passed as it did with the "_URL" class attribute.

It is possible to disable the url formatting using optional argument
"format_url".