Fix #1 - Problems uploading media
There are two problems uploading media to GCS:

1- wrongUrlForUpload:

Google's API documentation [1] now indicates that there are two
different URLs for uploads, one for metadata requests and another one
for the media upload requests.

This patch updates the URL following yarcherny's suggestion from the bug
report fixing the issue.

2- request.put doesn't seem to accept "bytearray" structure anymore, so
we'll be passing a "memoryview" instead.  We also change from
"bytes((data)" to "data.tobytes()" in "GCSObjFile" because it doesn't
work properly on Python 2.7.

[1]: https://cloud.google.com/storage/docs/json_api/v1/objects/insert